pipe tracking

+++ backend/app/models/stock.py
@@ -13,7 +13,7 @@ class StockLedger(Base):
     transaction_type = Column(String(30), nullable=False)
     reference_id = Column(Integer)
     reference_type = Column(String(30))
-    quantity = Column(Integer, nullable=False)
+    quantity = Column(Numeric(10, 3), nullable=False)
     unit_cost = Column(Numeric(10, 2))
     transaction_date = Column(Date, nullable=False)
     reason = Column(String(200), nullable=True)
